project/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-yasg
python-dotenv
factory_boy
black
